move plotters to plot module
