+++ src/hps_align/plot/derivative_plots.py
